+++ projects/igniteui-angular/src/lib/slider/thumb/thumb-slider.component.ts
@@ -57,13 +57,13 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     public deactiveState: boolean;
 
     @Output()
-    public onThumbValueChange = new EventEmitter<number>();
+    public thumbValueChange = new EventEmitter<number>();
 
     @Output()
     public onChange = new EventEmitter<any>();
 
     @Output()
-    public onHoverChange = new EventEmitter<boolean>();
+    public hoverChange = new EventEmitter<boolean>();
 
     @HostBinding('attr.tabindex')
     public tabindex = 0;
@@ -120,12 +120,12 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
 
     @HostListener('pointerenter')
     public onPinterEnter() {
-        this.onHoverChange.emit(true);
+        this.hoverChange.emit(true);
     }
 
     @HostListener('pointerleave')
     public onPointerLeave() {
-        this.onHoverChange.emit(false);
+        this.hoverChange.emit(false);
     }
 
     @HostListener('keydown', ['$event'])
@@ -144,7 +144,7 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
         }
 
         this.onChange.emit();
-        this.onThumbValueChange.emit(increment);
+        this.thumbValueChange.emit(increment);
     }
 
     @HostListener('blur')
@@ -195,7 +195,7 @@ export class IgxSliderThumbComponent implements OnInit, OnDestroy {
     private updateThumbValue(mouseX: number) {
         const updateValue = this.calculateTrackUpdate(mouseX);
         if (this.isActive && updateValue !== 0) {
-            this.onThumbValueChange.emit(updateValue);
+            this.thumbValueChange.emit(updateValue);
         }
     }